+++ golem-worker-service-base/src/api_definition/api_common.rs
@@ -5,7 +5,7 @@ use bincode::{Decode, Encode};
 use poem_openapi::NewType;
 use serde::{Deserialize, Serialize};
 
-use crate::worker_binding::golem_worker_binding::GolemWorkerBinding;
+use crate::worker_binding::GolemWorkerBinding;
 
 // Common to API definitions regardless of different protocols
 #[derive(Debug, Clone, Eq, PartialEq, Hash, Serialize, Deserialize, Encode, Decode, NewType)]
