+++ golem-worker-service-base/src/worker_binding/mod.rs
@@ -1,2 +1,5 @@
-pub(crate) mod golem_worker_binding;
-pub(crate) mod worker_binding_resolver;
+pub(crate) use golem_worker_binding::*;
+pub(crate) use worker_binding_resolver::*;
+
+mod golem_worker_binding;
+mod worker_binding_resolver;
