+++ golem-worker-service-base/src/worker_binding/worker_binding_resolver.rs
@@ -1,6 +1,6 @@
 use golem_wasm_rpc::TypeAnnotatedValue;
 
-use crate::worker_binding::golem_worker_binding::GolemWorkerBinding;
+use crate::worker_binding::GolemWorkerBinding;
 
 // For any input request type, there should be a way to resolve the
 // worker binding template, which is then used to form the worker request
