+++ golem-worker-service/src/api/register_api_definition_api.rs
@@ -9,10 +9,10 @@ use tracing::{error, info};
 use golem_service_base::api_tags::ApiTags;
 use golem_worker_service_base::api::ApiEndpointError;
 use golem_worker_service_base::api::HttpApiDefinition;
-use golem_worker_service_base::auth::{CommonNamespace, EmptyAuthCtx};
-use golem_worker_service_base::api_definition::{ApiDefinitionId, ApiVersion};
 use golem_worker_service_base::api_definition::http::get_api_definition_from_oas;
 use golem_worker_service_base::api_definition::http::HttpApiDefinition as CoreHttpApiDefinition;
+use golem_worker_service_base::api_definition::{ApiDefinitionId, ApiVersion};
+use golem_worker_service_base::auth::{CommonNamespace, EmptyAuthCtx};
 use golem_worker_service_base::service::api_definition::ApiDefinitionService;
 use golem_worker_service_base::service::http::http_api_definition_validator::RouteValidationError;
